pyproject.toml
[project]
name = "phzipcodes-web"
version = "0.1.0"
description = "Philippines Zip Codes Lookup Portal"
authors = [{ name = "Jayson", email = "jsoncp@proton.me" }]
license = "MIT"
readme = "README.md"

requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.6",
    "jinja2>=3.1.5",
    "mypy>=1.14.1",
    "phzipcodes>=0.1.6",
    "pytest>=8.3.4",
    "ruff>=0.8.4",
    "uvicorn>=0.34.0",
    "pytest-asyncio>=0.23.5",
    "pytest-playwright>=0.4.3",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "-n auto"
